Finalizadas pruebas.

+++ src/main/java/org/springframework/samples/petclinic/web/DiscountController.java
@@ -55,7 +55,7 @@ public class DiscountController {
 	@PostMapping(value = "/discounts/new")
 	public String processNewDiscountForm(@Valid Discount discount, BindingResult result, Product product,
 			@PathVariable("shopId") int shopId) {
-		if(discount.getFinishDate() != null && discount.getStartDate() != null) {
+		if (discount.getFinishDate() != null && discount.getStartDate() != null) {
 			if (discount.getFinishDate().isBefore(discount.getStartDate())) {
 				result.rejectValue("finishDate", "wrongDate", "Finish date must be after than start date");
 			}
@@ -94,15 +94,20 @@ public class DiscountController {
 			return "redirect:/shops/" + shopId + "/products/{productId}";
 		}
 	}
-	
+
 	@GetMapping("/discounts/{discountId}/delete")
-	public String processDeleteDiscount(@PathVariable("discountId") int discountId, Product product, @PathVariable("shopId") int shopId) {
+	public String processDeleteDiscount(@PathVariable("discountId") int discountId, Product product,
+			@PathVariable("shopId") int shopId) {
 		Discount discount = discountService.findDiscountById(discountId);
-		if(product.getDiscount().equals(discount)) {
-			product.setDiscount(null);
-			productService.saveProduct(product);
-			discountService.deleteDiscount(discount);
-			return "redirect:/shops/" + shopId + "/products/{productId}";
+		if (product.getDiscount() != null) {
+			if (product.getDiscount().equals(discount)) {
+				product.setDiscount(null);
+				productService.saveProduct(product);
+				discountService.deleteDiscount(discount);
+				return "redirect:/shops/" + shopId + "/products/{productId}";
+			} else {
+				return "/exception";
+			}
 		} else {
 			return "/exception";
 		}
